Change style files' name
