+++ src/components/BackDrop.js
@@ -1,11 +1,11 @@
-import backdropStyle from '../styles/components/backdrop.js';
+import backdropStyle from '../styles/components/backdropStyle.js';
 
 class Backdrop extends HTMLElement {
   constructor() {
     super();
 
     this.attachShadow({ mode: 'open' });
-    this.shadowRoot.append(backdropStyle());
+    this.shadowRoot.append(backdropStyle.call(this));
   }
 }
