+++ src/components/Sidebar/SidebarSub/SidebarBack.js
@@ -1,4 +1,4 @@
-import sidebarBackStyle from '../../../styles/components/sidebar/sidebarSub/sidebarBack.js';
+import sidebarBackStyle from '../../../styles/components/sidebar/sidebarSub/sidebarBackStyle.js';
 
 class SideBarBack extends HTMLElement {
   constructor() {
@@ -6,7 +6,7 @@ class SideBarBack extends HTMLElement {
 
     const shadow = this.attachShadow({ mode: 'open' });
     shadow.innerHTML = `<button><icon-element name="arrow-left" fill="var(--gray-600)"></icon-element>주메뉴</button>`;
-    this.shadowRoot.append(sidebarBackStyle());
+    this.shadowRoot.append(sidebarBackStyle.call(this));
   }
 }
