+++ src/styles/components/backdropStyle.js
