+++ src/styles/components/imageStyle.js
