+++ src/styles/components/navbar/mainMenuStyle.js
