+++ src/styles/components/sidebar/sidebarCategoryStyle.js
