+++ src/styles/components/sidebar/sidebarHeaderStyle.js
