+++ src/styles/components/sidebar/sidebarTitleStyle.js
